refactor: solver/tangent を simulation/ へ移動、stress_update を result.py にリネーム
core/solver.py と core/tangent.py は _PythonIntegratorBase 専用の実装詳細だったため
simulation/solver.py・simulation/tangent.py へ移動。Fortran integrator はこれらを
使わないため、Python integrator の内部として simulation 側に置く方が層構造に一致する。

core/stress_update.py は関数が存在せず dataclass のみのため result.py にリネームし、
ファイル名と内容の乖離を解消。公開 API (manforge.core.{ReturnMappingResult,StressUpdateResult})
は変更なし。CLAUDE.md・README.md・docstring の stale 記述も同時に修正。
