+++ src/manforge/core/result.py
